+++ src/CodeSharing.WebPortal/Controllers/Components/CanvasMenuViewComponent.cs
@@ -1,7 +1,7 @@
 using System.Security.Claims;
 using CodeSharing.WebPortal.Extensions;
 using CodeSharing.WebPortal.Interfaces;
-using CodeSharing.WebPortal.Models;
+using CodeSharing.WebPortal.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 
 namespace CodeSharing.WebPortal.Controllers.Components;
