+++ src/CodeSharing.WebPortal/Controllers/ContactController.cs
@@ -1,5 +1,5 @@
 using CodeSharing.WebPortal.Interfaces;
-using CodeSharing.WebPortal.Models;
+using CodeSharing.WebPortal.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 
 namespace CodeSharing.WebPortal.Controllers;
