+++ src/CodeSharing.WebPortal/ViewModels/ContactViewModel.cs
@@ -1,6 +1,6 @@
 using CodeSharing.ViewModels.Contents.Contact;
 
-namespace CodeSharing.WebPortal.Models;
+namespace CodeSharing.WebPortal.ViewModels;
 
 public class ContactViewModel
 {
