+++ src/CodeSharing.WebPortal/ViewModels/ListByCategoryIdViewModel.cs
@@ -2,7 +2,7 @@ using CodeSharing.Utilities.Commons;
 using CodeSharing.ViewModels.Contents.Category;
 using CodeSharing.ViewModels.Contents.Post;
 
-namespace CodeSharing.WebPortal.Models;
+namespace CodeSharing.WebPortal.ViewModels;
 
 public class ListByCategoryIdViewModel
 {
