+++ src/CodeSharing.WebPortal/ViewModels/ListByTagIdViewModel.cs
@@ -2,7 +2,7 @@ using CodeSharing.Utilities.Commons;
 using CodeSharing.ViewModels.Contents.Label;
 using CodeSharing.ViewModels.Contents.Post;
 
-namespace CodeSharing.WebPortal.Models;
+namespace CodeSharing.WebPortal.ViewModels;
 
 public class ListByTagIdViewModel
 {
