+++ src/CodeSharing.WebPortal/ViewModels/SearchPostViewModel.cs
@@ -1,7 +1,7 @@
 using CodeSharing.Utilities.Commons;
 using CodeSharing.ViewModels.Contents.Post;
 
-namespace CodeSharing.WebPortal.Models;
+namespace CodeSharing.WebPortal.ViewModels;
 
 public class SearchPostViewModel
 {
