+++ src/CodeSharing.WebPortal/Views/Contact/Index.cshtml
@@ -1,4 +1,4 @@
-@model ContactViewModel
+@model CodeSharing.WebPortal.ViewModels.ContactViewModel
 
 @{
     Layout = "_SubLayout";
