+++ src/CodeSharing.WebPortal/Views/Home/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model HomeViewModel
+﻿@model CodeSharing.WebPortal.ViewModels.HomeViewModel
 
 @{
   ViewData["Title"] = "Code Sharing · Chia sẻ mọi thứ bạn cần";
