+++ src/CodeSharing.WebPortal/Views/Post/Details.cshtml
@@ -1,4 +1,4 @@
-@model PostDetailViewModel
+@model CodeSharing.WebPortal.ViewModels.PostDetailViewModel
 
 @{
     Layout = "_SubLayout";
