+++ src/CodeSharing.WebPortal/Views/Post/ListByCategoryId.cshtml
@@ -1,4 +1,4 @@
-@model ListByCategoryIdViewModel
+@model CodeSharing.WebPortal.ViewModels.ListByCategoryIdViewModel
 
 @{
     Layout = "_Layout";
