+++ src/CodeSharing.WebPortal/Views/Post/ListByTagId.cshtml
@@ -1,4 +1,4 @@
-@model ListByTagIdViewModel
+@model CodeSharing.WebPortal.ViewModels.ListByTagIdViewModel
 
 @{
     Layout = "_Layout";
