+++ src/CodeSharing.WebPortal/Views/Post/Search.cshtml
@@ -1,4 +1,4 @@
-@model SearchPostViewModel
+@model CodeSharing.WebPortal.ViewModels.SearchPostViewModel
 
 @{
     Layout = "_Layout";
