+++ src/CodeSharing.WebPortal/Views/Shared/Components/CanvasMenu/Default.cshtml
@@ -1,4 +1,4 @@
-@model CanvasMenuViewModel
+@model CodeSharing.WebPortal.ViewModels.CanvasMenuViewModel
 
 @if (User.Identity!.IsAuthenticated)
 {
