+++ src/CodeSharing.WebPortal/Views/Shared/Components/Header/Default.cshtml
@@ -1,4 +1,4 @@
-@model HeaderViewModel
+@model CodeSharing.WebPortal.ViewModels.HeaderViewModel
 
 <!-- header -->
 <header class="header-personal">
