+++ src/CodeSharing.WebPortal/Views/Shared/Components/SideBar/Default.cshtml
@@ -1,4 +1,4 @@
-@model SideBarViewModel
+@model CodeSharing.WebPortal.ViewModels.SideBarViewModel
 
 <div class="sidebar">
     <!-- widget about -->
