+++ src/CodeSharing.WebPortal/Views/Shared/Error.cshtml
@@ -1,4 +1,4 @@
-@model ErrorViewModel
+@model CodeSharing.WebPortal.ViewModels.ErrorViewModel
 @{
     ViewData["Title"] = "Error";
 }
