+++ src/CodeSharing.WebPortal/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
 ﻿@using CodeSharing.WebPortal
-@using CodeSharing.WebPortal.Models
+@using CodeSharing.WebPortal.ViewModels
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
