+++ src/CodeSharing.WebPortal/appsettings.json
@@ -7,5 +7,10 @@
   },
   "PageSize": 8,
   "PageIndexSize": 10,
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "RecaptchaOption": {
+    "SiteKey": "6LfDqL4fAAAAAFr0HkqwZvYPlRvwHzzmUB1KscC5",
+    "SecretKey": "6LfDqL4fAAAAALWgKbe5BWf2zi4oMZMMhpRAM_c0",
+    "Url" :  "https://www.google.com/recaptcha/api/siteverify?"
+  }
 }
